requirements-dev.txt
-r requirements.txt

pytest
psycopg2-binary
pytest-cov
pytest-django
flake8
ipython
parameterized
black
isort
django-extensions
django-cors-headers
mixer
coverage
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
bandit==1.7.0
liccheck==0.5.3
safety
